requirements.txt
pycryptodome
pandas
numpy
matplotlib
